Moved the main threading components to their own folder
